+++ Changelog.md
@@ -24,6 +24,8 @@
 * Modify Sonic Spear damage and take enemy's size into account
 * Turn the gelatinous cube into a passive monster
 * Adapt how statues react when being hit to the original Larn
+* Scrollable windows can be scrolled with the mouse wheen or by clicking
+  the arrows
 * Allow dragging windows with the mouse
 
 ### Fixed bugs